[DEV-1521] Updated LNbits config, renamed Bicep variable, refactoring
