+++ .vscode/settings.json
@@ -21,6 +21,8 @@
     "instrumentationkey",
     "kauf",
     "kucoin",
+    "lnurl",
+    "lnurlp",
     "masternode",
     "masternodes",
     "meshsdk",
