+++ infrastructure/config/lightning/lnbits/dev.env
@@ -1,6 +1,6 @@
 LNBITS_BASEURL=https://vm-dfx-btc-inp-dev.westeurope.cloudapp.azure.com
 
-HOST=127.0.0.1
+HOST=0.0.0.0
 PORT=5000
 
 FORWARDED_ALLOW_IPS="*"
