+++ src/integration/lightning/lnurl-forward.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { HttpService } from 'src/shared/services/http.service';
-import { LnurlPayRequestDto } from './dto/lnurlp-payrequest.dto';
+import { LnurlPayRequestDto } from './dto/lnurlp-pay-request.dto';
 import { LightningHelper } from './lightning-helper';
 
 @Injectable()
@@ -8,8 +8,8 @@ export class LnUrlForwardService {
   constructor(private readonly http: HttpService) {}
 
   async lnurlpForward(id: string): Promise<LnurlPayRequestDto> {
-    const lnbitsUrl = `${LightningHelper.lnurlpLnBitsBasePath}/${id}`;
-    const payRequest = await this.http.get<LnurlPayRequestDto>(lnbitsUrl);
+    const lnBitsUrl = `${LightningHelper.lnurlpLnBitsBasePath}/${id}`;
+    const payRequest = await this.http.get<LnurlPayRequestDto>(lnBitsUrl);
 
     payRequest.callback = `${LightningHelper.lnurlpCallbackDfxApiBasePath}/${id}`;
     payRequest.metadata = '[["text/plain", "DFX Deposit Address"]]';
@@ -17,8 +17,8 @@ export class LnUrlForwardService {
     return payRequest;
   }
 
-  async lnurlpCallbackForward(id: string, amount: number): Promise<any> {
-    const lnbitsCallbackUrl = `${LightningHelper.lnurlpCallbackLnBitsBasePath}/${id}?amount=${amount}`;
-    return this.http.get<any>(lnbitsCallbackUrl);
+  async lnurlpCallbackForward(id: string, params: any): Promise<any> {
+    const lnBitsCallbackUrl = `${LightningHelper.lnurlpCallbackLnBitsBasePath}/${id}`;
+    return this.http.get<any>(lnBitsCallbackUrl, { params });
   }
 }
